+++ src/services/fetchers/index.ts
@@ -6,6 +6,8 @@ import PolygonIOPriceFetcher from './PolygonIOPriceFetcher';
 import CryptoComparePriceWSFetcher from './CryptoComparePriceWSFetcher';
 import IEXEnergyFetcher from './IEXEnergyFetcher';
 import CoinmarketcapPriceFetcher from './CoinmarketcapPriceFetcher';
+import CoinmarketcapHistoHourFetcher from './CoinmarketcapHistoHourFetcher';
+import CoinmarketcapHistoDayFetcher from './CoinmarketcapHistoDayFetcher';
 import CoingeckoPriceFetcher from './CoingeckoPriceFetcher';
 import BEACPIAverageFetcher from './BEACPIAverageFetcher';
 import OnChainDataFetcher from './OnChainDataFetcher';
@@ -20,6 +22,8 @@ export {
   IEXEnergyFetcher,
   CoingeckoPriceFetcher,
   CoinmarketcapPriceFetcher,
+  CoinmarketcapHistoHourFetcher,
+  CoinmarketcapHistoDayFetcher,
   BEACPIAverageFetcher,
   OnChainDataFetcher,
 };