+++ src/services/FeedProcessor.ts
@@ -43,6 +43,9 @@ class FeedProcessor {
     @inject(fetchers.IEXEnergyFetcher) IEXEnergyFetcher: fetchers.IEXEnergyFetcher,
     @inject(fetchers.CoingeckoPriceFetcher) CoingeckoPriceFetcher: fetchers.CoingeckoPriceFetcher,
     @inject(fetchers.CoinmarketcapPriceFetcher) CoinmarketcapPriceFetcher: fetchers.CoinmarketcapPriceFetcher,
+    @inject(fetchers.CoinmarketcapHistoHourFetcher)
+    CoinmarketcapHistoHourFetcher: fetchers.CoinmarketcapHistoHourFetcher,
+    @inject(fetchers.CoinmarketcapHistoDayFetcher) CoinmarketcapHistoDayFetcher: fetchers.CoinmarketcapHistoDayFetcher,
     @inject(fetchers.BEACPIAverageFetcher) BEACPIAverageFetcher: fetchers.BEACPIAverageFetcher,
     @inject(fetchers.OnChainDataFetcher) OnChainDataFetcher: fetchers.OnChainDataFetcher,
   ) {
@@ -55,6 +58,8 @@ class FeedProcessor {
       IEXEnergyFetcher,
       CoingeckoPriceFetcher,
       CoinmarketcapPriceFetcher,
+      CoinmarketcapHistoHourFetcher,
+      CoinmarketcapHistoDayFetcher,
       BEACPIAverageFetcher,
       OnChainDataFetcher,
     };