+++ src/services/fetchers/PolygonIOPriceFetcher.ts
@@ -6,6 +6,7 @@ import PolygonIOPriceService from '../PolygonIOPriceService';
 class PolygonIOPriceFetcher {
   @inject(PolygonIOPriceService) polygonIOPriceService!: PolygonIOPriceService;
 
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types,@typescript-eslint/no-explicit-any
   async apply({sym}: any, timestamp: number): Promise<number> {
     const price = await this.polygonIOPriceService.getLatestPrice(sym, timestamp);
     if (price !== null) {