+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- new fetcher CoinmarketcapHistoDay
+- new fetcher CoinmarketcapHistoHour
 
 ## [4.0.0] - 2021-06-29
 ### Added