+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.1.0] - 2021-07-02
+### Added
+- new fetcher CoinmarketcapHistoDay
+- new fetcher CoinmarketcapHistoHour
+
 ## [4.0.0] - 2021-06-29
 ### Added
 - `OnChain` feed collector